pyproject.toml
[project]
name = "proyecto2-cifrados"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi~=25.1.0",
    "cryptography~=44.0.2",
    "email-validator>=2.2.0",
    "fastapi~=0.115.8",
    "fastapi-mail>=1.5.0",
    "google-auth>=2.40.2",
    "httpx>=0.28.1",
    "passlib==1.7.4",
    "protobuf>=6.31.1",
    "pydantic~=2.10.6",
    "pyjwt~=2.10.1",
    "python-dotenv>=1.1.0",
    "python-jose==3.4.0",
    "python-multipart~=0.0.20",
    "requests>=2.32.3",
    "sqlalchemy~=2.0.38",
    "uvicorn~=0.34.0",
]
